docs: Improve, simplify Storybook config & update to 6.2.9
- Correct build
- Move theme-editor stories to theme-editor package directory
- Centralize storybook configuration in storybook/.storybook
- Remove ControlColors story (not useful)
- Fix fast mode
- Update storybook (6.2.9)
