+++ storybook/package.json
@@ -12,19 +12,19 @@
     "@looker/components": "^1.4.0",
     "@looker/components-theme-editor": "^1.4.0",
     "@looker/design-tokens": "^1.3.1",
-    "@storybook/react": "^6.1.15",
+    "@storybook/react": "^6.2.9",
     "lodash": "^4.17.20",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-is": "^16.13.1",
-    "storybook": "^6.1.15",
+    "storybook": "^6.2.9",
     "styled-components": "5.3.0"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "^6.1.15",
-    "@storybook/addon-essentials": "^6.1.15",
-    "@storybook/addon-storyshots": "^6.1.15",
-    "@storybook/addon-storyshots-puppeteer": "^6.1.15",
+    "@storybook/addon-a11y": "^6.2.9",
+    "@storybook/addon-essentials": "^6.2.9",
+    "@storybook/addon-storyshots": "^6.2.9",
+    "@storybook/addon-storyshots-puppeteer": "^6.2.9",
     "babel-loader-exclude-node-modules-except": "^1.1.2",
     "tsconfig-paths-webpack-plugin": "^3.5.1"
   }
